Major consolidation update & official support for OS X App Bundles
	* Goodbye POSIX build option! This has been phased out entirely.
The default nwo is to set CMAKE_INSTALL_PREFIX to the current build dir
only if the user has not modified this variable, otherwise use their
preset and our ending-paths (we adhere to the POSIX hierarchy if we are
not building an OS X app bundle). This will probably make more sense
when I update comments later.
	* Consolidated installation targets; setting variables and passing
once to the install command helps clean a lot of this up.
	* Introduce macros.cmake with install_rpath macro
	* Source files list have been relocated to src/CMakeLists.txt