Introduce wide-scale usage of GameObject pointer & bug fixes
	* This pointer reference is intended for holding objects that can be
considered "global" or nearly so, granting us ease of access between
game states. Side benefit of greatly improving performance by only
loading things once!

	* Fix bug where the opposing player's selected card is not reset
entirely between states; this is introduced in the clear method of
CardHand.
	- Introduce CardHand::exists & CardHand::randomize
		* Ensure that the player never has two of the same card as well
as never an invalid ID; something buggy is happening with the seed
generator every so many times -- possibly to do with the integer size?
	* Introduce App::onInit method
	* Commenting out audio subsystem until a couple bugs are fixed...