Refactor CPU player (complete rewrite)
Today marks a very special day for TTcards! An AI quite capable of
whooping your ass has been implemented. Although only the base game
rules have been implemented in this draft, this doesn't stop the AI
from playing with any rule-set enabled.

* PlayState: All of the AI logic now resides in the CPU_Player class.
This state is now responsible just for initializing the CPU player.

* CardHand: Fix ::shuffle method's lack of respect for passed level_min
and level_max arguments.

* BoardTile: Introduce ::set_bounds && equality, inequality operators
for the class. This is used to compliment the introduction of
BoardTile::null, used in the CPU_Player class.