Overhaul of the time step game loop
This simplifies the time step logic in the game -- doing only the
absolute minimum we need at the moment. In short, we rely on either
the refresh rate of the display determining our frame rate, or we
run in a fixed time step, capping the frame rate.
  The DELAY_TIME configuration var is no longer necessary due to
these changes and has been removed.

* Add VSYNC, FRAME_RATE configuration vars for tweaking the time
stepping of the game.
* Add VIDEO_DISPLAY_INDEX configuration var for the setting game
window up on a specific display index.