Update xcode build scripts to use nomlib's improved ver
This gives us a few new optional features: a) passing of the build
configuration type, defaulting to Debug; b) passing of the num of
threads xcodebuild is allowed to use; c) bit of additoinal status
output of what the shell script is doing.

* bin/xcode_install.sh: Add failure output when this script opts
out of trying to do anything. Additionally, the exit code for
this script under said failure condition has been changed from
always returning 1 to now match the exit code of the failed command,
for sake of clarity.